isolate dense output further
